% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emojis_reference}
\alias{emojis_reference}
\title{Emojis reference data}
\format{
A data frame with 1566 rows and 5 variables:
\describe{
\item{\code{id}}{Emoji ID}
\item{\code{name}}{Name}
\item{\code{version}}{Version number}
\item{\code{keywords}}{Keywords}
\item{\code{alias}}{Alias}
}
}
\source{
\url{https://raw.githubusercontent.com/missive/emoji-mart/main/packages/emoji-mart-data/sets/5/native.json}
}
\usage{
emojis_reference
}
\description{
A simplified data frame with a list of emojis available in Felt through the
Emoji Mart picker. More information: \url{https://github.com/missive/emoji-mart}
}
\keyword{datasets}
