% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{felt_delete_map_layer}
\alias{felt_delete_map_layer}
\title{Delete Layer from an Existing Map}
\usage{
felt_delete_map_layer(map_id, layer_id)
}
\arguments{
\item{map_id}{map identifier from url, from \verb{https://felt.com/map/Readable-Name-map_id}}

\item{layer_id}{layer identifier, as returned by \code{felt_get_map_layers()} or \code{felt_add_map_layers()}}
}
\value{
response code
}
\description{
Delete Layer from an Existing Map
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
layer <- felt_add_map_layers('TBI8sDkmQjuK2GX9CSiHiUA', 'test', fs::path_package('felt', 'towns.geojson'))
felt_delete_map_layers(map_id = 'Rockland-2024-Districts-TBI8sDkmQjuK2GX9CSiHiUA', layer)
\dontshow{\}) # examplesIf}
}
