% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finish_upload.R
\name{felt_finish_upload}
\alias{felt_finish_upload}
\title{Finish Upload}
\usage{
felt_finish_upload(map_id)
}
\arguments{
\item{map_id}{map identifier from url, from \verb{https://felt.com/map/Readable-Name-map_id}}
}
\value{
\code{httr2} status code
}
\description{
Uploads are redirected to Amazon S3 buckets and must be confirmed once done.
This function is \emph{primarily} internal, but may have uses for custom uploads.
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
felt_finish_upload('TBI8sDkmQjuK2GX9CSiHiUA')
\dontshow{\}) # examplesIf}
}
\concept{edit}
