% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{felt_get_library}
\alias{felt_get_library}
\title{List all layers in your workspace library}
\usage{
felt_get_library(source = "workspace", clean = TRUE)
}
\arguments{
\item{source}{character, source of the layers, one of \code{workspace}, \code{felt}, or \code{all}}

\item{clean}{logical, whether to turn the API response into a \link[tibble:tibble]{tibble::tibble}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} if \code{clean = TRUE}, otherwise a list
}
\description{
List all layers in your workspace library
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
felt_get_library()
\dontshow{\}) # examplesIf}
}
