% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{felt_get_map}
\alias{felt_get_map}
\title{Get Map Information from Map ID}
\usage{
felt_get_map(map_id)
}
\arguments{
\item{map_id}{map identifier from url, from \verb{https://felt.com/map/Map-ID}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} for the map
}
\description{
Get Map Information from Map ID
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
felt_get_map('Phoenix-Arizona-L9Cefd3GJSGuYQaU5tBOlvA')
\dontshow{\}) # examplesIf}
}
