% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_elements.R
\name{felt_get_map_elements}
\alias{felt_get_map_elements}
\title{Get Map Elements from Map ID}
\usage{
felt_get_map_elements(map_id)
}
\arguments{
\item{map_id}{map identifier from url, from \verb{https://felt.com/map/Map-ID}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} for the map
}
\description{
Get Map Elements from Map ID
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
felt_get_map_elements(map_id = 'Rockland-2024-Districts-TBI8sDkmQjuK2GX9CSiHiUA')
\dontshow{\}) # examplesIf}
}
