% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{felt_get_style}
\alias{felt_get_style}
\title{Get Felt Style Information for a Layer}
\usage{
felt_get_style(map_id, layer_id)
}
\arguments{
\item{map_id}{map identifier from url, from \verb{https://felt.com/map/Readable-Name-map_id}}

\item{layer_id}{layer identifier, as returned by \code{felt_get_map_layers()} or \code{felt_add_map_layers()}}
}
\value{
a list version of style information
}
\description{
For details on the Felt Style Language, see \url{https://felt.com/blog/felt-style-language}.
}
\examples{
felt_get_style(map_id = 'TBI8sDkmQjuK2GX9CSiHiUA', layer_id = 'Qno0gfmkTS9BggbesfQooND')
}
\concept{style}
