% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_user.R
\name{felt_user}
\alias{felt_user}
\title{Get the user information associated with the default (or supplied) token}
\usage{
felt_user(token = NULL)
}
\arguments{
\item{token}{Felt personal access token}
}
\description{
List the name, email address, and user ID for the Felt user associated with
the default (or supplied) token.
}
