% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feltr-package.R
\docType{package}
\name{feltr-package}
\alias{feltr}
\alias{feltr-package}
\title{feltr: Access the Felt API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Upload, download, and edit internet maps with the Felt API (\url{https://developers.felt.com/rest-api/getting-started}). Allows users to create new maps, edit existing maps, and extract data. Provides tools for working with layers, which represent geographic data, and elements, which are interactive annotations. Spatial data accessed from the API is transformed to work with 'sf'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/feltr}
  \item \url{https://christophertkenny.com/feltr/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
