% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_comments.R
\name{get_felt_comments}
\alias{get_felt_comments}
\title{Get comments from a Felt map}
\usage{
get_felt_comments(
  map_id,
  simplifyVector = TRUE,
  geometry = TRUE,
  crs = NULL,
  token = NULL
)
}
\arguments{
\item{map_id}{A Felt map URL or map ID.}

\item{simplifyVector}{Passed to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json()}}, Default: \code{TRUE}}

\item{geometry}{If \code{TRUE}, return a \code{sf} object. Default: \code{TRUE}}

\item{crs}{Coordinated reference system to return (if geometry is TRUE),
Default: \code{NULL}}

\item{token}{Felt personal access token}
}
\value{
A data frame or simple feature object with a list column of comments.
}
\description{
Get comments from a Felt map
}
\details{
API documentation
\url{https://feltmaps.notion.site/Felt-Public-API-reference-c01e0e6b0d954a678c608131b894e8e1#0d5e58ee84e0445d8484445a27be1d48}
}
