% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_styles.R
\name{get_felt_style}
\alias{get_felt_style}
\alias{felt_layer_styles}
\alias{update_felt_style}
\title{Get Felt layer styles or update a layer style}
\usage{
get_felt_style(map_id, layer_id = NULL, call = caller_env())

update_felt_style(map_id, style, layer_id = NULL, call = caller_env())
}
\arguments{
\item{map_id}{A Felt map URL or map ID.}

\item{layer_id}{If \code{NULL} (default), all layers for the map are used.
Multi-layer maps are not currently supported. Otherwise use a layer ID
string. Use \code{\link[=read_felt_layers]{read_felt_layers()}} to list layers for an existing map.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{style}{A named list that can be converted to a valid Felt Style
Language string. If style is supplied with a datasets id value matching the
layer datasets ids, this function updates an existing layer style. If style
is \code{NULL} (default), read styles for supplied map and layer. See the
\href{https://feltmaps.notion.site/Felt-Style-Language-de08179cb8494d3d8bbf5fb970f03fd0}{documentation on the Felt Style Language}
and the \href{https://feltmaps.notion.site/Felt-Public-API-reference-c01e0e6b0d954a678c608131b894e8e1#722105ec74e7492cb934bf81338db8b5}{API endpoint for updating layer styles}
for more information.}
}
\value{
If layer_id is \code{NULL} and the map contains multiple styles or if
layer_id is a character vector, the function returns a list with style
elements named with the layer ID values. If layer_id is a string, the
function returns a named list with a single Felt Style Language
specification.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Get one or more Felt layer styles or update a specified layer style. Warning,
updating a layer style without a list that can be converted to a valid Felt
Style Language (FSL) may get a layer into an \emph{irreversible broken state}.
}
