% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_felt_url}
\alias{is_felt_url}
\alias{check_felt_url}
\title{Check Felt URLs}
\usage{
is_felt_url(x)

check_felt_url(
  url,
  message = "{.arg url} must be a Felt URL.",
  call = caller_env(),
  ...
)
}
\description{
Check Felt URLs
}
\keyword{internal}
