% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_felt.R
\name{read_felt}
\alias{read_felt}
\title{Read data from a Felt map}
\usage{
read_felt(
  url,
  map_id = NULL,
  ...,
  crs = 3857,
  token = NULL,
  rename = TRUE,
  name_repair = "check_unique"
)
}
\arguments{
\item{map_id}{A Felt map URL, map ID string, or a named list with a id and
type element. If map_id is a list, it must be equivalent to the output from
\code{\link[=get_felt_map]{get_felt_map()}} where the list includes a "id" string and a "type" string
with the value "map".}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}.}

\item{crs}{Coordinate reference system to return. Defaults to 3857.}

\item{token}{Felt personal access token}

\item{rename}{If \code{TRUE} (default), strip the prefix text "felt-" from all
column names.}

\item{name_repair}{Passed to repair parameter of \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
Defaults to "check_unique".}
}
\value{
A simple feature data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Read simple features from a Felt map or get data embedded in the website of a
Felt map. Superseded by \code{\link[=read_felt_map]{read_felt_map()}}.
}
\seealso{
\code{\link[sf:st_read]{sf::read_sf()}}
}
