% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_map.R
\name{read_felt_map}
\alias{read_felt_map}
\alias{create_felt_map}
\alias{delete_felt_map}
\title{Read, create, or delete a Felt map}
\usage{
read_felt_map(url, map_id = NULL, ..., token = NULL)

create_felt_map(
  title = NULL,
  description = NULL,
  location = NULL,
  zoom = NULL,
  layer_urls = NULL,
  basemap = c("default", "satellite"),
  token = NULL
)

delete_felt_map(url, map_id = NULL, safely = TRUE, token = NULL)
}
\arguments{
\item{url}{A Felt map URL.}

\item{map_id}{A Felt map ID, optional if url is supplied.}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}.}

\item{token}{Felt personal access token}

\item{title}{Map title}

\item{description}{Map description}

\item{location}{Location to center map, either a sf or sfc object or a length
2 numeric vector in the form of \code{c("lon", "lat")}}

\item{zoom}{Zoom level number}

\item{layer_urls}{Raster layer URLs}

\item{basemap}{Basemap, string ("default" or "satellite") or a valid layer
URL or color as hex code.}

\item{safely}{If \code{TRUE} (default), check for user confirmation before
deleting a Felt map. If \code{FALSE}, delete map without checking for
confirmation.}
}
\value{
\code{\link[=read_felt_map]{read_felt_map()}} returns a sf object, \code{\link[=create_felt_map]{create_felt_map()}} invisibly
returns a list of attributes for the created map, and \code{\link[=delete_felt_map]{delete_felt_map()}}
does not return anything.
}
\description{
Read, create, or delete a Felt map from a URL or map ID.
}
