% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_felt_raster.R
\name{read_felt_raster}
\alias{read_felt_raster}
\title{Use rasterpic to create a \code{SpatRaster} object from a Felt map}
\usage{
read_felt_raster(x, images, ..., crs = 3857)
}
\arguments{
\item{x}{If x is a Felt map URL, it is passed to \code{\link[=read_felt]{read_felt()}} to create a
data.frame of features with a "type" and "text" columns. If x is a
data.frame, it is expected to be a data.frame created by reading a Felt map
with \code{\link[=read_felt]{read_felt()}} but could be a sf object with a type column that
includes the value "Image" and (if images is named) a text column with
matching text.}

\item{images}{A vector of image file paths or URLs with a "png", "jpeg/jpg",
or "tiff/tif" file extension. images must be ordered to match the order of
"Image" type features in the input data.frame x or have names that match
the text column for x. If images is named, any "Image" features in x with
text that does not match the names for images are excluded from the
returned list.}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}.}

\item{crs}{Coordinate reference system to return. Defaults to 3857.}
}
\value{
If images is length 1, a \code{SpatRaster} object is returned. Otherwise,
the function returns a list of \code{SpatRaster} objects of the same length as
images.
}
\description{
Read an image feature from Felt and use the \code{\link[rasterpic:rasterpic_img]{rasterpic::rasterpic_img()}}
function and a corresponding image URL or file path to create a \code{SpatRaster}
object based on the feature geometry.
}
