% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_felt.R
\name{request_felt}
\alias{request_felt}
\title{Build and perform a request for the Felt API}
\usage{
request_felt(
  base_url = "https://felt.com/api/v1",
  endpoint = NULL,
  token = NULL,
  ...,
  perform = TRUE,
  call = caller_env()
)
}
\arguments{
\item{base_url}{Base URL for request.}

\item{token}{Felt personal access token}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Build and perform a request for the Felt API
}
\keyword{internal}
