% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_map.R
\name{set_felt_map_data}
\alias{set_felt_map_data}
\title{Set up Felt map data}
\usage{
set_felt_map_data(
  title = NULL,
  description = NULL,
  location = NULL,
  zoom = NULL,
  layer_urls = NULL,
  basemap = c("default", "satellite"),
  allow_null = TRUE,
  call = caller_env()
)
}
\arguments{
\item{title}{Map title}

\item{description}{Map description}

\item{location}{Location to center map, either a sf, sfc, or bbox object or a
length 2 numeric vector in the form of \code{c("lon", "lat")}. To pass
coordinates in lat/lon order, set the \code{feltr.latlon} option to \code{TRUE}
(option defaults to \code{FALSE}). If location is \code{NULL} (default), map is
centered on Oakland, California.}

\item{zoom}{Zoom level number}

\item{layer_urls}{A character vector or list of raster layer URLs.}

\item{basemap}{Basemap, string ("default" or "satellite"), a valid layer URL,
a color name, or a color hex code.}
}
\description{
Internal function to prepare input parameters for \code{\link[=create_felt_map]{create_felt_map()}}
}
\keyword{internal}
