% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_felt_token.R
\name{set_felt_token}
\alias{set_felt_token}
\alias{get_felt_token}
\title{Set or get a Felt API personal access token}
\usage{
set_felt_token(
  token = NULL,
  install = FALSE,
  overwrite = FALSE,
  default = "FELT_ACCESS_TOKEN"
)

get_felt_token(
  token = NULL,
  default = "FELT_ACCESS_TOKEN",
  call = caller_env()
)
}
\arguments{
\item{token}{Felt personal access token}

\item{install}{Add your token to your \code{.Renviron} for use in future sessions.}

\item{overwrite}{If \code{TRUE}, overwrite any existing token in \code{.Renviron},}

\item{default}{Default name used for environmental variable where the token
is stored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
An API personal access token is required to use \code{\link[=read_felt]{read_felt()}}. See
\url{https://feltmaps.notion.site/Felt-Public-API-reference-PUBLIC-c01e0e6b0d954a678c608131b894e8e1}
for instructions on how to get a token.
}
