% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_layers.R
\name{create_felt_layer}
\alias{create_felt_layer}
\alias{delete_felt_layer}
\alias{read_felt_layers}
\alias{update_felt_layer}
\title{Read layers from a Felt map, delete a layer, or create a new layer}
\usage{
create_felt_layer(
  map_id,
  layer,
  name = NULL,
  fileext = "gpkg",
  ...,
  fill_color = NULL,
  stroke_color = NULL,
  webhook_url = NULL,
  token = NULL
)

delete_felt_layer(map_id, layer_id = NULL, safely = TRUE, token = NULL)

read_felt_layers(
  map_id,
  simplifyVector = TRUE,
  token = NULL,
  call = caller_env()
)

update_felt_layer(
  map_id,
  layer_id,
  name = NULL,
  description = NULL,
  simplifyVector = TRUE,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{map_id}{A Felt map URL, map ID string, or a named list with a id and
type element. If map_id is a list, it must be equivalent to the output from
\code{\link[=get_felt_map]{get_felt_map()}} where the list includes a "id" string and a "type" string
with the value "map".}

\item{layer}{Required. A object, file path, a layer source URL, or a \code{sf} or
\code{sfc} object. If layer is a file path or a source URL, the file type or URL
type must be supported by Felt. See
\url{https://feltmaps.notion.site/Upload-Anything-b26d739e80184127872faa923b55d232#3e37f06bc38c4971b435fbff2f4da6cb}
for details. If layer is a \code{sf} or \code{sfc} object, the object is saved to a
temporary file using the supplied fileext.}

\item{name}{Name for new map layer.}

\item{fileext}{File extension to use for temporary file if layer is a \code{sf} or
\code{sfc} object.}

\item{...}{Additional parameters passed to \code{\link[sf:st_write]{sf::st_write()}} if layer is a
\code{sf} or \code{sfc} object.}

\item{fill_color, stroke_color}{Hex string to use as the layer fill or stroke
color. Optional.}

\item{webhook_url}{When the layer finishes processing, Felt will notify to
this URL.}

\item{token}{Felt personal access token}

\item{layer_id}{Layer ID. Layer IDs for a map can be listed using
\code{\link[=read_felt_layers]{read_felt_layers()}}}

\item{safely}{If \code{TRUE} (default), check for user confirmation before
deleting a Felt map. If \code{FALSE}, delete map without checking.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{description}{Map description}
}
\description{
Read layers from a Felt map with \code{\link[=read_felt_layers]{read_felt_layers()}}, delete a single layer
with \code{\link[=delete_felt_layer]{delete_felt_layer()}}, update a layer with \code{\link[=update_felt_layer]{update_felt_layer()}}, or
create a new layer from a URL, file, or sf or sfc object with
\code{\link[=create_felt_layer]{create_felt_layer()}}. Note that reading layers does not return layer
data—only a list of layers.
}
