% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_layers.R
\name{create_felt_layer}
\alias{create_felt_layer}
\alias{delete_felt_layer}
\alias{read_felt_layers}
\title{Read layers from a Felt map, delete a layer, or create a new layer from a URL}
\usage{
create_felt_layer(map_id, layer, name = NULL, token = NULL)

delete_felt_layer(map_id, layer_id = NULL, safely = TRUE, token = NULL)

read_felt_layers(
  map_id,
  simplifyVector = TRUE,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{map_id}{A Felt map URL or map ID.}

\item{layer}{Required. Layer source URL, either a supported static file or a
URL supported by Felt. See
\url{https://feltmaps.notion.site/Upload-Anything-b26d739e80184127872faa923b55d232#3e37f06bc38c4971b435fbff2f4da6cb}}

\item{name}{Name for new map layer.}

\item{token}{Felt personal access token}

\item{layer_id}{Layer ID. Layer IDs for a map can be listed using
\code{\link[=read_felt_layers]{read_felt_layers()}}}

\item{safely}{If \code{TRUE} (default), check for user confirmation before
deleting a Felt map. If \code{FALSE}, delete map without checking.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Read layers from a Felt map, delete a single layer, or create a new layer
from a URL. Note that the layers do not include data.
}
\examples{

create_felt_layer()

}
