% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_layers.R
\name{create_felt_layer}
\alias{create_felt_layer}
\alias{delete_felt_layer}
\alias{read_felt_layers}
\title{Read layers from a Felt map, delete a layer, or create a new layer}
\usage{
create_felt_layer(
  map_id,
  layer,
  name = NULL,
  fileext = "gpkg",
  ...,
  token = NULL
)

delete_felt_layer(map_id, layer_id = NULL, safely = TRUE, token = NULL)

read_felt_layers(
  map_id,
  simplifyVector = TRUE,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{map_id}{A Felt map URL or map ID.}

\item{layer}{Required. A object, file path, a layer source URL, or a \code{sf} or
\code{sfc} object. If layer is a file path or a source URL, the file type or URL
type must be supported by Felt. See
\url{https://feltmaps.notion.site/Upload-Anything-b26d739e80184127872faa923b55d232#3e37f06bc38c4971b435fbff2f4da6cb}
for details. If layer is a \code{sf} or \code{sfc} object, the object is saved to a
temporary file using the supplied fileext.}

\item{name}{Name for new map layer.}

\item{fileext}{File extension to use for temporary file if layer is a \code{sf} or
\code{sfc} object.}

\item{...}{Additional parameters passed to \code{\link[sf:st_write]{sf::st_write()}} if layer is a
\code{sf} or \code{sfc} object.}

\item{token}{Felt personal access token}

\item{layer_id}{Layer ID. Layer IDs for a map can be listed using
\code{\link[=read_felt_layers]{read_felt_layers()}}}

\item{safely}{If \code{TRUE} (default), check for user confirmation before
deleting a Felt map. If \code{FALSE}, delete map without checking.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Read layers from a Felt map, delete a single layer, or create a new layer
from a URL, file, or sf or sfc object. Note that reading layers does not
return layer data—only a list of layers.
}
\examples{

create_felt_layer()

}
