% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{felt_create_map}
\alias{felt_create_map}
\title{Create a new map}
\usage{
felt_create_map(
  title = NULL,
  basemap = NULL,
  layer_urls = NULL,
  lat = NULL,
  lon = NULL,
  zoom = NULL
)
}
\arguments{
\item{title}{Title to use for the map. Defaults to \code{NULL}.}

\item{basemap}{Basemap for the new map. Defaults to \code{NULL}. Can be a URL or color hex code}

\item{layer_urls}{vector of URLs to generate layers in map. Defaults to \code{NULL}.}

\item{lat}{latitude to center the map. Defaults to \code{NULL}.}

\item{lon}{longitude to center the map. Defaults to \code{NULL}}

\item{zoom}{zoom level to initialize the map with.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} for the new map
}
\description{
Create a new map
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
map <- felt_create_map(title = 'feltr example')
map
# and delete it again
felt_delete_map(map_id = map$id)
\dontshow{\}) # examplesIf}
}
\concept{map}
