% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{felt_get_comments}
\alias{felt_get_comments}
\title{Export Comments on an Existing Map}
\usage{
felt_get_comments(map_id, clean = TRUE)
}
\arguments{
\item{map_id}{character, map identifier from url, from \verb{https://felt.com/map/Readable-Name-map_id}}

\item{clean}{logical, whether to turn the API response into a \link[tibble:tibble]{tibble::tibble}}
}
\value{
a \link[tibble:tibble]{tibble::tibble} for the map
}
\description{
Export Comments on an Existing Map
}
\examples{
\dontshow{if (has_felt_key()) withAutoprint(\{ # examplesIf}
felt_get_comments(map_id = 'Rockland-2024-Districts-TBI8sDkmQjuK2GX9CSiHiUA')
\dontshow{\}) # examplesIf}
}
\concept{comments}
