% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{felt_get_user}
\alias{felt_get_user}
\title{Obtain information about the user}
\usage{
felt_get_user()
}
\value{
a \link[tibble:tibble]{tibble::tibble} of information about the user
}
\description{
Obtain information about the user
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
felt_get_user()
\dontshow{\}) # examplesIf}
}
