% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{felt_patch_style}
\alias{felt_patch_style}
\title{Update Felt Style Information for a Layer}
\usage{
felt_patch_style(map_id, layer_id, fsl)
}
\arguments{
\item{map_id}{map identifier from url, from \verb{https://felt.com/map/Readable-Name-map_id}}

\item{layer_id}{layer identifier, as returned by \code{felt_get_map_layers()} or \code{felt_add_map_layers()}}

\item{fsl}{A list indicating the Felt style language to update the layer to}
}
\value{
response data
}
\description{
For details on the Felt Style Language, see \url{https://felt.com/blog/felt-style-language}.
}
\examples{
\dontshow{if (has_felt_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
felt_get_style(map_id = 'TBI8sDkmQjuK2GX9CSiHiUA', layer_id = 'Qno0gfmkTS9BggbesfQooND')
\dontshow{\}) # examplesIf}
}
\concept{style}
