% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feltr-package.R
\docType{package}
\name{feltr-package}
\alias{feltr}
\alias{feltr-package}
\title{feltr: Read Spatial Data from Felt}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Read Felt maps to simple feature objects in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/feltr/}
  \item \url{https://elipousson.github.io/feltr/}
  \item Report bugs at \url{https://github.com/elipousson/feltr/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
