% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_felt_url}
\alias{is_felt_url}
\alias{check_felt_url}
\title{Check Felt URLs}
\usage{
is_felt_url(x)

check_felt_url(
  url,
  message = "{.arg url} must be a Felt URL.",
  call = caller_env(),
  ...
)
}
\arguments{
\item{x, url}{Object to check.}

\item{message}{Message passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} if url is not a Felt url.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Check Felt URLs
}
