% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_url.R
\name{is_felt_url}
\alias{is_felt_url}
\alias{check_felt_url}
\title{Is a object a Felt URL?}
\usage{
is_felt_url(x)

check_felt_url(x, allow_null = FALSE, arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{Object to check.}

\item{allow_null}{If \code{TRUE}, \code{\link[=check_felt_url]{check_felt_url()}} allows a \code{NULL} input without
an error. Defaults to \code{FALSE}.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Is a object a Felt URL?
}
