% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_felt.R
\name{read_felt}
\alias{read_felt}
\title{Read data from a Felt map URL}
\usage{
read_felt(url, ..., type = "features")
}
\arguments{
\item{url}{A Felt URL}

\item{...}{Additional parameters passed to [sf::read_sf()]}
}
\value{
A simple feature data.frame.
}
\description{
Read data from a Felt map URL as a GeoJSON.
}
\seealso{
[sf::read_sf()]
}
