% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_felt.R
\name{read_felt}
\alias{read_felt}
\alias{get_felt_data}
\title{Read data from a Felt map}
\usage{
read_felt(
  url,
  map_id = NULL,
  ...,
  crs = 3857,
  token = NULL,
  rename = TRUE,
  name_repair = "check_unique"
)

get_felt_data(url, call = caller_env())
}
\arguments{
\item{url}{A Felt map URL.}

\item{map_id}{A Felt map ID, optional if url is supplied.}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}.}

\item{crs}{Coordinate reference system to return. Defaults to 3857.}

\item{token}{Felt personal access token}

\item{rename}{If \code{TRUE} (default), strip the prefix text "felt-" from all
column names.}

\item{name_repair}{Passed to repair parameter of \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
Defaults to "check_unique".}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A simple feature data frame or a list of the parsed JSON found
in the "felt-data" div of a Felt map webpage.
}
\description{
Read simple features from a Felt map or get data embedded in the website of a
Felt map. \code{\link[=read_felt]{read_felt()}} is a wrapper for \code{\link[=read_felt_map]{read_felt_map()}} with a few extra
features. \code{\link[=get_felt_data]{get_felt_data()}} returns the parsed JSON included in the body of a
Felt map website (which includes both features and other user and layer
metadata).
}
\seealso{
\code{\link[sf:st_read]{sf::read_sf()}}
}
