% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_felt.R
\name{read_felt}
\alias{read_felt}
\title{Read data from a Felt map URL}
\usage{
read_felt(url, ..., type = "features", rename = TRUE)
}
\arguments{
\item{url}{A Felt URL}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}}

\item{type}{Type of data to return. Defaults to "features".}

\item{rename}{If \code{TRUE} (default), strip the prefix text "felt-" from all
column names.}
}
\value{
A simple feature data.frame.
}
\description{
Read data from a Felt map URL as a GeoJSON.
}
\seealso{
\code{\link[sf:st_read]{sf::read_sf()}}
}
