% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felt_map.R
\name{set_felt_map_data}
\alias{set_felt_map_data}
\title{Set up Felt map data}
\usage{
set_felt_map_data(
  title = NULL,
  description = NULL,
  location = NULL,
  zoom = NULL,
  layer_urls = NULL,
  basemap = c("default", "satellite"),
  call = caller_env()
)
}
\arguments{
\item{title}{Map title}

\item{description}{Map description}

\item{location}{Location to center map, either a sf or sfc object or a length
2 numeric vector in the form of \code{c("lon", "lat")}}

\item{zoom}{Zoom level number}

\item{layer_urls}{Raster layer URLs}

\item{basemap}{Basemap, string ("default" or "satellite") or a valid layer
URL or color as hex code.}
}
\description{
Internal function for setting up map data for \code{\link[=create_felt_map]{create_felt_map()}}
}
\keyword{internal}
